+++ lib/src/statements/statements.dart
@@ -1,9 +1,11 @@
+export 'add_column.dart';
 export 'and.dart';
 export 'column.dart';
 export 'column_definition.dart';
 export 'comparisons.dart';
 export 'create_table.dart';
 export 'delete.dart';
+export 'drop_column.dart';
 export 'drop_table.dart';
 export 'equals.dart';
 export 'group.dart';